requirements.txt
python-dotenv
telethon
socketserver
threading
http